requirements.txt
# Core
Django==5.2.9
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.11

# REST / API
djangorestframework==3.15.2
django-filter==24.3

# Swagger / OpenAPI
drf-spectacular==0.27.2

# Excel export
openpyxl==3.1.5
